central_controller/requirements.txt
flask>=3.0.0
pika>=1.3.2
openai>=1.0.0
diffusers>=0.20.0
torch>=2.1.0
minio>=7.2.0
tensorflow>=2.15.0
opencv-python>=4.9.0
moviepy>=1.0.3
pydub>=0.25.1
google-api-python-client>=2.100.0
tenacity>=8.2.3
transformers>=4.35.0
python-dotenv>=1.0.0
prometheus_client>=0.20.0